fix: Update files and testing for Settings rename
