+++ client/test/cypress/integration/settings.cy.js
@@ -1,22 +1,32 @@
 /// <reference types="Cypress" />
 /// <reference path="../support/index.d.ts" />
 
-import { a11yLogViolations } from '../support/helpers'
+import { a11yLogViolations } from "../support/helpers"
 
-describe("Login and Password", () => {
+describe("Settings page", () => {
   beforeEach(() => {
     cy.task("resetDb")
     cy.login({ email: "regularuser@pilcrow.dev" })
-    cy.visit("/account/login")
+    cy.visit("/account/settings")
+  })
+
+  it("can update the email field", () => {
+    cy.injectAxe()
+    cy.dataCy("update_user_email").clear().type("updateduser@pilcrow.dev")
+    cy.dataCy("button_save").click()
+    cy.dataCy("update_user_notify")
+      .should("be.visible")
+      .should("have.class", "bg-positive")
+    cy.checkA11y(null, null, a11yLogViolations)
   })
 
   it("can update the password field", () => {
+    cy.injectAxe()
     cy.dataCy("update_user_password").clear().type("XMYeygtC7TuxgER4")
     cy.dataCy("button_save").click()
     cy.dataCy("update_user_notify")
       .should("be.visible")
       .should("have.class", "bg-positive")
-    // Inject the axe-core libraray
   })
 
   it("should assert the page is accessible", () => {
@@ -25,11 +35,4 @@ describe("Login and Password", () => {
     cy.dataCy("vueAccount")
     cy.checkA11y(null, null, a11yLogViolations)
   })
-
-  // TODO: Uncomment once email updates work again
-  // it('can update the email field', () => {
-  //   cy.dataCy('update_user_email').clear().type('updateduser@pilcrow.dev');
-  //   cy.dataCy('update_user_button_save').click();
-  //   cy.dataCy('update_user_notify').should('be.visible').should('have.class','bg-positive');
-  // });
 })
